modules init function only called if not disabled
Before: `Skada:IsDisabled(...)` was used on the top of each module to make sure to skip the code if the it is disabled.
After: the function is called on Skada loading process and modules functions aren't called if they are disabled. However, since some module depend on others, the function is still used to make sure the dependency is not disabled.